[#30] refactor/dto 레이어 이동
