+++ src/main/java/com/anipick/backend/anime/controller/dto/MetaDataGroupResultDto.java
@@ -1,6 +1,9 @@
-package com.anipick.backend.anime.dto;
+package com.anipick.backend.anime.controller.dto;
 
 import com.anipick.backend.anime.domain.AnimeFormat;
+import com.anipick.backend.anime.service.dto.GenreDto;
+import com.anipick.backend.anime.service.dto.SeasonDto;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
