+++ src/main/java/com/anipick/backend/anime/mapper/MetaDataMapper.java
@@ -1,6 +1,6 @@
 package com.anipick.backend.anime.mapper;
 
-import com.anipick.backend.anime.dto.GenreDto;
+import com.anipick.backend.anime.service.dto.GenreDto;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
