+++ src/main/java/com/anipick/backend/anime/service/dto/GenreDto.java
@@ -1,4 +1,4 @@
-package com.anipick.backend.anime.dto;
+package com.anipick.backend.anime.service.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
