+++ src/main/java/com/anipick/backend/explore/service/ExploreService.java
@@ -7,9 +7,9 @@ import com.anipick.backend.anime.domain.SeasonConverter;
 import com.anipick.backend.common.domain.SortOption;
 import com.anipick.backend.common.dto.CursorDto;
 import com.anipick.backend.explore.domain.GenresOption;
-import com.anipick.backend.explore.dto.ExploreItemDto;
-import com.anipick.backend.explore.dto.ExplorePageDto;
-import com.anipick.backend.explore.dto.ExploreRequestDto;
+import com.anipick.backend.explore.service.dto.ExploreItemDto;
+import com.anipick.backend.explore.controller.dto.ExplorePageDto;
+import com.anipick.backend.explore.service.dto.ExploreRequestDto;
 import com.anipick.backend.explore.mapper.ExploreMapper;
 
 import lombok.RequiredArgsConstructor;
