+++ src/main/java/com/anipick/backend/explore/service/dto/ExploreRequestDto.java
@@ -1,4 +1,4 @@
-package com.anipick.backend.explore.dto;
+package com.anipick.backend.explore.service.dto;
 
 import com.anipick.backend.anime.domain.RangeDate;
 import lombok.AllArgsConstructor;
