+++ src/main/java/com/anipick/backend/review/service/ReviewService.java
@@ -1,8 +1,8 @@
 package com.anipick.backend.review.service;
 
 import com.anipick.backend.common.dto.CursorDto;
-import com.anipick.backend.review.dto.RecentReviewItemDto;
-import com.anipick.backend.review.dto.RecentReviewPageDto;
+import com.anipick.backend.review.service.dto.RecentReviewItemDto;
+import com.anipick.backend.review.controller.dto.RecentReviewPageDto;
 import com.anipick.backend.review.mapper.RecentReviewMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
