+++ src/main/java/com/anipick/backend/user/controller/UserController.java
@@ -2,7 +2,7 @@ package com.anipick.backend.user.controller;
 
 import com.anipick.backend.common.dto.ApiResponse;
 import com.anipick.backend.user.domain.User;
-import com.anipick.backend.user.dto.SignUpRequest;
+import com.anipick.backend.user.controller.dto.SignUpRequest;
 import com.anipick.backend.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
