+++ src/main/java/com/anipick/backend/user/controller/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.anipick.backend.user.dto;
+package com.anipick.backend.user.controller.dto;
 
 import lombok.Getter;
